requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
requests
psycopg2-binary
python-dotenv
SQLAlchemy
gunicorn
azure-identity
azure-kusto-data
pandas
azure-ai-projects 
azure-ai-inference
